--- app/Http/Controllers/GetWeekDcardController.php
@@ -1,114 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\GetWeekDcard;
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use mysql_xdevapi\Exception;
-
-class GetWeekDcardController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return JsonResponse
-     */
-    public function index(): JsonResponse
-    {
-        try {
-//            $day = date('w');
-//            $week_start = date('Y-m-d', strtotime('-'.$day.' days'));
-//            $week_end = date('Y-m-d', strtotime('+'.(6-$day).' days'));
-            $week_start = "2021-11-07";
-            $week_end = "2021-11-13";
-            $dcardAll = DB::table('dcard_rawdata')
-                ->leftJoin('nlp_analysis', 'dcard_rawdata.Id', '=', 'nlp_analysis.Id')
-                ->leftJoin('comparison', 'comparison.Id', '=', 'nlp_analysis.Id')
-                ->select('dcard_rawdata.Id', 'dcard_rawdata.Title', 'dcard_rawdata.CreatedAt', 'dcard_rawdata.Content'
-                    , 'nlp_analysis.SA_Score', 'nlp_analysis.SA_Class', 'comparison.Level', 'comparison.KeywordLevel1',
-                    'comparison.KeywordLevel2', 'comparison.KeywordLevel3')
-                ->whereBetween('dcard_rawdata.CreatedAt', [$week_start, $week_end])
-                ->orderByDesc('dcard_rawdata.Id')
-                ->get();
-        } catch (Exception $e) {
-            $error['message'] = '404 Not Found!!' . $e;
-            return response()->json($error, 404);
-        } finally {
-            if (!$dcardAll->isEmpty()){
-                return response()->json($dcardAll, 200, ['Content-Type' => 'application/json;charset=UTF-8', 'Charset' => 'utf-8'],
-                    JSON_UNESCAPED_UNICODE);
-            } else {
-                $error['message'] = '404 Not Found!!';
-                return response()->json($error, 404);
-            }
-        }
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\GetWeekDcard  $getWeekDcard
-     * @return \Illuminate\Http\Response
-     */
-    public function show(GetWeekDcard $getWeekDcard)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\GetWeekDcard  $getWeekDcard
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(GetWeekDcard $getWeekDcard)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\GetWeekDcard  $getWeekDcard
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, GetWeekDcard $getWeekDcard)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\GetWeekDcard  $getWeekDcard
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(GetWeekDcard $getWeekDcard)
-    {
-        //
-    }
-}
