--- app/Models/GetAllDcard.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Laravel\Scout\Searchable;
-use Spatie\QueryBuilder\QueryBuilder;
-
-
-class GetAllDcard extends Model
-{
-    use HasFactory, Searchable;
-
-    public function toSearchableArray()
-    {
-        $array = $this->toArray();
-
-        // Customize the data array...
-
-        return $array;
-    }
-}
