--- app/Models/GetArticleId.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class GetArticleId extends Model
-{
-    use HasFactory;
-}
